[REFACTOR][IR] Migrate include/tvm/node into include/tvm/ir
The node/ directory was a legacy artifact from when runtime "node"
infrastructure was separate from IR-level concerns.  All remaining
public headers have direct IR semantics and belong under tvm/ir/.

Moves (header + include-guard rename):
- include/tvm/node/functor.h       → include/tvm/ir/node_functor.h
- include/tvm/node/cast.h          → include/tvm/ir/cast.h
- include/tvm/node/repr.h          → include/tvm/ir/repr.h
- include/tvm/node/script_printer.h→ include/tvm/ir/script_printer.h
- include/tvm/node/attr_registry_map.h → include/tvm/ir/attr_registry_map.h

Source moves:
- src/node/repr.cc          → src/ir/repr.cc
- src/node/script_printer.cc→ src/ir/script_printer.cc

Deletes (dead shims / empty stubs):
- include/tvm/node/repr_printer.h     (shim → tvm/node/repr.h)
- include/tvm/node/structural_equal.h (shim → tvm/ffi/extra/structural_equal.h)
- include/tvm/node/structural_hash.h  (shim → tvm/ffi/extra/structural_hash.h)
- src/node/repr_printer.cc            (empty stub)
- src/node/container_printing.cc      (empty stub)

All ~35 in-tree consumers updated to use the new paths.
CMakeLists.txt: drop the now-empty src/node/*.cc glob.
