+++ include/tvm/ir/analysis.h
@@ -30,7 +30,7 @@
 #include <tvm/ffi/container/array.h>
 #include <tvm/ir/expr.h>
 #include <tvm/ir/module.h>
-#include <tvm/node/functor.h>
+#include <tvm/ir/node_functor.h>
 
 namespace tvm {
 namespace ir {
