+++ include/tvm/ir/repr.h
@@ -17,18 +17,17 @@
  * under the License.
  */
 /*!
- * \file tvm/node/repr.h
+ * \file tvm/ir/repr.h
  * \brief ostream operator<< for ObjectRef, Any, and Variant, delegating to
  *        ffi::ReprPrint.  Also re-exports the Dump() debug helpers.
  *
  * Include this header wherever you need `os << some_objectref` and you are
  * no longer pulling in the legacy repr_printer.h.
  */
-#ifndef TVM_NODE_REPR_H_
-#define TVM_NODE_REPR_H_
+#ifndef TVM_IR_REPR_H_
+#define TVM_IR_REPR_H_
 
 #include <tvm/ffi/extra/dataclass.h>
-#include <tvm/ffi/reflection/access_path.h>
 #include <tvm/runtime/object.h>
 
 #include <iostream>
@@ -68,51 +67,6 @@ inline std::ostream& operator<<(std::ostream& os, const ffi::Variant<V...>& n) {
   return os << ffi::ReprPrint(Any(n));
 }
 
-namespace reflection {
-
-inline std::ostream& operator<<(std::ostream& os, const AccessStep& step) {
-  namespace refl = ffi::reflection;
-  switch (step->kind) {
-    case refl::AccessKind::kAttr: {
-      os << '.' << step->key.cast<ffi::String>();
-      return os;
-    }
-    case refl::AccessKind::kArrayItem: {
-      os << "[" << step->key.cast<int64_t>() << "]";
-      return os;
-    }
-    case refl::AccessKind::kMapItem: {
-      os << "[" << step->key << "]";
-      return os;
-    }
-    case refl::AccessKind::kAttrMissing: {
-      os << ".<missing attr " << step->key.cast<ffi::String>() << "`>";
-      return os;
-    }
-    case refl::AccessKind::kArrayItemMissing: {
-      os << "[<missing item at " << step->key.cast<int64_t>() << ">]";
-      return os;
-    }
-    case refl::AccessKind::kMapItemMissing: {
-      os << "[<missing item at " << step->key << ">]";
-      return os;
-    }
-    default: {
-      TVM_FFI_THROW(InternalError) << "Unknown access step kind: " << static_cast<int>(step->kind);
-    }
-  }
-  return os;
-}
-
-inline std::ostream& operator<<(std::ostream& os, const AccessPath& path) {
-  ffi::Array<AccessStep> steps = path->ToSteps();
-  os << "<root>";
-  for (const auto& step : steps) {
-    os << step;
-  }
-  return os;
-}
-}  // namespace reflection
 }  // namespace ffi
 }  // namespace tvm
-#endif  // TVM_NODE_REPR_H_
+#endif  // TVM_IR_REPR_H_
