+++ include/tvm/tirx/buffer.h
@@ -28,7 +28,7 @@
 #include <tvm/ffi/reflection/registry.h>
 #include <tvm/ffi/string.h>
 #include <tvm/ir/expr.h>
-#include <tvm/node/script_printer.h>
+#include <tvm/ir/script_printer.h>
 #include <tvm/tirx/var.h>
 
 #include <string>
