+++ include/tvm/tirx/expr.h
@@ -29,7 +29,7 @@
 #include <tvm/ffi/container/map.h>
 #include <tvm/ffi/string.h>
 #include <tvm/ir/expr.h>
-#include <tvm/node/functor.h>
+#include <tvm/ir/node_functor.h>
 #include <tvm/runtime/base.h>
 #include <tvm/runtime/data_type.h>
 #include <tvm/tirx/buffer.h>
