+++ src/ir/attr_functor.h
@@ -30,7 +30,7 @@
 #ifndef TVM_IR_ATTR_FUNCTOR_H_
 #define TVM_IR_ATTR_FUNCTOR_H_
 
-#include <tvm/node/functor.h>
+#include <tvm/ir/node_functor.h>
 #include <tvm/tirx/expr.h>
 
 #include <utility>
