--- src/node/container_printing.cc
@@ -1,26 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-/*!
- * \file node/container_printing.cc
- * \brief DEPRECATED — tvm-ffi provides built-in repr for Array/Map/Shape.
- *        The legacy ReprPrinter dispatches for containers are no longer needed.
- */
-// This file is intentionally empty. Container repr is handled by
-// ffi::ReprPrint (tvm-ffi/src/ffi/extra/dataclass.cc).
