+++ tests/cpp/arith_simplify_test.cc
@@ -19,7 +19,7 @@
 
 #include <gtest/gtest.h>
 #include <tvm/arith/analyzer.h>
-#include <tvm/node/structural_equal.h>
+#include <tvm/ffi/extra/structural_equal.h>
 #include <tvm/runtime/logging.h>
 #include <tvm/te/operation.h>
