+++ tests/cpp/nested_msg_test.cc
@@ -18,7 +18,7 @@
  */
 
 #include <gtest/gtest.h>
-#include <tvm/node/structural_equal.h>
+#include <tvm/ffi/extra/structural_equal.h>
 #include <tvm/relax/nested_msg.h>
 #include <tvm/relax/struct_info.h>
 #include <tvm/runtime/data_type.h>
